refactor: 글 작성 페이지의 로직을 hooks로 분리하여 관심사 분리 수행
