+++ src/app/write/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import dynamic from 'next/dynamic';
-import Image from 'next/image';
 import { useState } from 'react';
 import '@blocknote/core/fonts/inter.css';
 import '@blocknote/react/style.css';
@@ -10,92 +9,24 @@ import {
   Input,
   NavigationIcons,
   Pencil,
-  StickerClose,
   StickerStar,
   Switch,
 } from '@/components/Common';
 import { usePortal } from '@/hooks';
-import { cn, generateId, safeSeesionStorage } from '@/utils';
-import { EDITOR_KEY } from '@/components/Common/Editor/constants';
-import { useToastContext } from '@/components/Common/Toast/ToastProvider';
-import { Ghost, Github, Smile, Star } from './dummyIcons';
-import { useRecoilStickerState } from './store';
-import { ImageProps } from './components/Sticker/types';
-import { usePost } from './api/queries';
+import StickerPannel from './components/StickerPannel';
+import { useWritePost } from './hooks';
 
 const Konva = dynamic(() => import('./components/Konva'), { ssr: false });
 const Editor = dynamic(() => import('@/components/Common/Editor'), {
   ssr: false,
 });
 
-// TODO: Icon이 선택되는 경우 색상이 변경되도록 attribute 설정
-const stickerTabs = [
-  { id: 0, Icon: Github },
-  { id: 1, Icon: Ghost },
-  { id: 2, Icon: Smile },
-  { id: 3, Icon: Star },
-] as const;
-
-const dummyImages = Array.from({ length: 12 }).reduce(
-  (arr: Array<{ id: number }>, _, idx) => [...arr, { id: idx + 1 }],
-  [],
-);
-
 export default function Page() {
   const [title, setTitle] = useState<string>('');
-  const [selectedId, setSelectedId] = useState<number>(0);
   const [editable, setEditable] = useState<boolean>(true);
 
   const port = usePortal({ id: 'write-portal-container' });
-  const { setStickerStates } = useRecoilStickerState();
-  const { handleError } = useToastContext()!;
-
-  // TODO: blogId 사용
-  const { mutate } = usePost(8);
-
-  const addStickerToPanel = ({
-    src,
-    width,
-    height,
-    x,
-    y,
-  }: Omit<ImageProps, 'resetButtonRef' | 'id'>) => {
-    setStickerStates((currentImages) => [
-      ...currentImages,
-      {
-        id: generateId(),
-        width,
-        height,
-        x,
-        y,
-        src,
-      },
-    ]);
-  };
-
-  // TODO: refactor
-  const handleSubmitPost = (publishState: 'publish' | 'save' = 'save') => {
-    const content = safeSeesionStorage.get(EDITOR_KEY) || '';
-
-    if (!content) {
-      handleError('내용을 입력해주세요', { duration: 2000 });
-      return;
-    }
-
-    if (!title) {
-      handleError('제목을 입력해주세요', { duration: 2000 });
-      return;
-    }
-
-    // TODO: blogId
-    mutate({
-      blogId: 10,
-      title,
-      content,
-      temporaryState: publishState === 'publish',
-      categoryName: '',
-    });
-  };
+  const { handleSubmitPost } = useWritePost();
 
   return (
     <section className="relative flex justify-center">
@@ -121,13 +52,13 @@ export default function Page() {
             {/* TODO: 글 업로드 */}
             <Button
               className="bg-[#E3E3E3] w-60 h-30"
-              onClick={() => handleSubmitPost()}
+              onClick={() => handleSubmitPost(title)}
             >
               저장
             </Button>
             <Button
               className="bg-primary text-[white] w-60 h-30"
-              onClick={() => handleSubmitPost('publish')}
+              onClick={() => handleSubmitPost(title, 'publish')}
             >
               발행
             </Button>
@@ -135,86 +66,7 @@ export default function Page() {
         </div>,
       )}
 
-      <aside
-        className={cn(
-          'w-300 h-[577px] rounded-16 px-23 py-20 absolute top-85 left-30 transition-all duration-200',
-          !editable && 'z-[210000001] shadow-background',
-        )}
-      >
-        <div className="flex justify-between">
-          <Button
-            onClick={() => {
-              setEditable((prev) => !prev);
-            }}
-            className="select-none text-[26px] bg-transparent font-luckiest !text-primary text-shadow-primary transition-all duration-200"
-          >
-            stickers
-          </Button>
-
-          {!editable && (
-            <Button
-              onClick={() => setEditable(() => true)}
-              className="bg-transparent"
-            >
-              <StickerClose />
-            </Button>
-          )}
-        </div>
-
-        {/* TODO: 검색 기능 구현 */}
-        {!editable && (
-          <Input
-            wrapperClassName="w-full mt-17"
-            className="rounded-12 border-1 border-[#D8D8D8] px-17 placeholder:text-[#999]"
-            placeholder="검색어 입력"
-          />
-        )}
-
-        {!editable && (
-          <section className="mt-60">
-            <div>
-              <div className="flex gap-2 px-20 mb-6">
-                {stickerTabs.map(({ id, Icon }) => (
-                  <Icon
-                    key={id}
-                    className={cn(
-                      'w-34 h-34 p-5 cursor-pointer',
-                      id === selectedId && 'bg-[#FFEBDF] rounded-4',
-                    )}
-                    onClick={() => setSelectedId(id)}
-                    selected={id === selectedId}
-                  />
-                ))}
-              </div>
-
-              <div className="w-full h-1 bg-[#D8D8D8]" />
-            </div>
-
-            <div className="grid grid-cols-3 justify-items-center items-center gap-20 mt-20">
-              {dummyImages.map(({ id }) => (
-                <Image
-                  key={`/images/stickers/sticker-${id}.svg`}
-                  src={`/images/stickers/sticker-${id}.svg`}
-                  alt={`/images/stickers/sticker-${id}.svg image`}
-                  width={60}
-                  height={60}
-                  className="cursor-pointer"
-                  onClick={() => {
-                    addStickerToPanel({
-                      src: `/images/stickers/sticker-${id}.svg`,
-                      width: 60,
-                      height: 60,
-                      x: 300,
-                      y: 300,
-                    });
-                    setEditable(() => true);
-                  }}
-                />
-              ))}
-            </div>
-          </section>
-        )}
-      </aside>
+      <StickerPannel editable={editable} setEditable={setEditable} />
 
       <Konva enable={!editable} />
